+++ tests/model/test_hazard_models.py
@@ -4,7 +4,6 @@ Basic model migration, structure
 
 import itertools
 
-import pandas as pd
 import pyarrow as pa
 import pyarrow.dataset as ds
 import pytest
@@ -12,9 +11,8 @@ from moto import mock_dynamodb
 from pyarrow import fs
 
 from toshi_hazard_store.model import drop_r4, migrate_r4
-
-# from toshi_hazard_store.model.hazard_models_pydantic import HazardAggregateCurve
-from toshi_hazard_store.model.pyarrow import pyarrow_dataset
+from toshi_hazard_store.model.hazard_models_pydantic import HazardAggregateCurve
+from toshi_hazard_store.model.pyarrow import pyarrow_aggr_dataset
 from toshi_hazard_store.model.revision_4 import hazard_aggregate_curve
 
 
@@ -154,13 +152,13 @@ class TestRevisionFourModelCreation_WithAdaption:
         assert res.sort_key == '-38.160~178.247:0250:PGA:mean:NSHM_DUMMY_MODEL'
 
 
-@pytest.fixture(scope='function')
+@pytest.fixture
 def pyarrow_aggregation_models(many_rlz_args):
-    def model_generator():
+    def generator_fn():
         for loc, vs30, imt, agg in itertools.product(
             many_rlz_args["locs"][:5], many_rlz_args["vs30s"], many_rlz_args["imts"], ['mean', 'cov', '0.95']
         ):
-            yield dict(
+            yield HazardAggregateCurve(
                 compatible_calc_id="NZSHM22",
                 hazard_model_id="MyNewModel",
                 nloc_001=loc.resample(0.001).code,
@@ -171,31 +169,70 @@ def pyarrow_aggregation_models(many_rlz_args):
                 values=[(x / 1000) for x in range(44)],
             )
 
-    df = pd.DataFrame(model_generator())
-    yield pa.Table.from_pandas(df)
+    yield generator_fn
 
 
-class TestPyarrowModelCreation:
+# Test pyarrow HazardAggregation models
+def test_HazardAggregation_roundtrip_dataset(pyarrow_aggregation_models, tmp_path):
 
-    # @pytest.mark.skip("Test needs schema, but do we still want this feature???")
-    def test_HazardAggregation_roundtrip_dataset(self, pyarrow_aggregation_models, tmp_path):
+    output_folder = tmp_path / "ds"
+
+    models = pyarrow_aggregation_models()
+
+    print(models)
+    filesystem = fs.LocalFileSystem()
+
+    # write the dataset
+    pyarrow_aggr_dataset.append_models_to_dataset(models, output_folder, filesystem=filesystem)
+
+    # read and check the dataset
+    dataset = ds.dataset(output_folder, filesystem=filesystem, format='parquet', partitioning='hive')
+    table = dataset.to_table()
+    df = table.to_pandas()
+
+    expected = pyarrow_aggr_dataset.table_from_models(pyarrow_aggregation_models())
+
+    assert table.shape == expected.shape
+    assert df.shape == expected.shape
+
+
+def test_HazardAggregation_write_dataset_with_bad_schema(pyarrow_aggregation_models, tmp_path, monkeypatch):
+
+    # monkeypatch with a bad schema
+    bad_schema = pa.schema([("mumbo", pa.string()), ("jumbo", pa.string())])
+    monkeypatch.setattr(pyarrow_aggr_dataset, "hazard_agreggate_schema", bad_schema)
+
+    # should raise excption about the schema mismatch
+    with pytest.raises(pa.ArrowTypeError, match=r"which does not match expected schema"):
 
         output_folder = tmp_path / "ds"
-
-        models = pyarrow_aggregation_models
-
-        print(models)
-
+        models = pyarrow_aggregation_models()
         filesystem = fs.LocalFileSystem()
 
-        # write the dataset
-        pyarrow_dataset.append_models_to_dataset(models, output_folder, filesystem=filesystem)
+        pyarrow_aggr_dataset.append_models_to_dataset(models, output_folder, filesystem=filesystem)
 
-        # read and check the dataset
-        dataset = ds.dataset(output_folder, filesystem=filesystem, format='parquet', partitioning='hive')
-        table = dataset.to_table()
-        df = table.to_pandas()
 
-        assert table.shape == models.shape
-        assert df.shape == models.shape
-        print(df)
+def test_HazardAggregation_read_dataset_with_bad_schema_doesnt_fail(pyarrow_aggregation_models, tmp_path, monkeypatch):
+
+    dataset_folder = tmp_path / "ds"
+    models = pyarrow_aggregation_models()
+    filesystem = fs.LocalFileSystem()
+
+    pyarrow_aggr_dataset.append_models_to_dataset(models, dataset_folder, filesystem=filesystem)
+
+    filesystem = fs.LocalFileSystem()
+    dataset = ds.dataset(dataset_folder, schema=None, format='parquet', filesystem=filesystem, partitioning='hive')
+
+    print()
+    print(dataset.schema)
+    print()
+
+    # I epected this should raise exception about the schema mismatch
+    # based on https://arrow.apache.org/docs/python/generated/pyarrow.\
+    #  dataset.Dataset.html#pyarrow.dataset.Dataset.replace_schema
+    bad_schema = pa.schema([("numeric", pa.int64()), ("mumbo", pa.string()), ("jumbo", pa.string())])
+    dataset = dataset.replace_schema(bad_schema)  # should raise exception but doesn't
+
+    table = dataset.to_table()
+    df = table.to_pandas()
+    print(df)  # still no exception